+++ node/src/lib/general-utils.js
